chore: main-app 사용되지 않는 파일 제거
사용되지 않는 리소스 파일 및 컴포넌트 파일 제거
app.js 관련 파일을 App폴더로 이동
