--- frontend/main-app/src/App.css
@@ -1,41 +1,0 @@
-/* .App {
-	text-align: center;
-}
-
-.App-logo {
-	height: 40vmin;
-}
-
-.App-header {
-	background-color: #282c34;
-	min-height: 100vh;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	font-size: calc(10px + 2vmin);
-	color: white;
-}
-
-.App-link {
-	color: #09d3ac;
-} */
-
-.App {
-	width: 100%;
-	height: 100%;
-	display: flex;
-	flex-direction: column;
-	color: #222222;
-}
-
-a {
-	color: #212121;
-	text-decoration: none;
-}
-
-a:hover {
-	color: #212121;
-	text-decoration: none;
-	cursor: pointer;
-}
