--- frontend/main-app/src/App.js
@@ -1,15 +1,0 @@
-import React, { useState } from "react";
-import "./App.css";
-import EventForm from "./components/EventForm";
-import HostLoginMessage from "./components/HostLoginMessage";
-
-function App() {
-	return (
-		<div className="App">
-			<EventForm />
-			<HostLoginMessage />
-		</div>
-	);
-}
-
-export default App;
