+++ frontend/main-app/src/App/App.css
@@ -1,0 +1,18 @@
+.App {
+	width: 100%;
+	height: 100%;
+	display: flex;
+	flex-direction: column;
+	color: #222222;
+}
+
+a {
+	color: #212121;
+	text-decoration: none;
+}
+
+a:hover {
+	color: #212121;
+	text-decoration: none;
+	cursor: pointer;
+}
